// src/app/_dto/iaexpsubgroup.dto.ts
// import { Iaexp } from "./Iaexp";
// import { Iaexpgroup } from "./Iaexpgroup";


export class IaexpsubgroupDto {

  coCode: string;

  groupCode: string;

  code: string;

  name: string;

  status: string;

  sysuser1: string | null;

  sysdate1: Date | null;

  modiuser: string | null;

  modidate: Date | null;

  catCode: string | null;

  // iaexps: Iaexp[];

  // iaexpgroup: Iaexpgroup;
}